RF: Switch from --skip-native to --output-space=...
* Use output spaces to select fsaverage subjects to copy to derivatives
* Update T1w and surface space logic in workflows/epi

+++ fmriprep/interfaces/bids.py
@@ -209,8 +209,9 @@ class BIDSFreeSurferDirInputSpec(BaseInterfaceInputSpec):
                                 desc='FreeSurfer installation directory')
     subjects_dir = traits.Str('freesurfer', usedefault=True,
                               desc='Name of FreeSurfer subjects directory')
+    spaces = traits.List(traits.Str, desc='Set of output spaces to prepare')
     overwrite_fsaverage = traits.Bool(False, usedefault=True,
-                                      desc='Overwrite fsaverage, if present')
+                                      desc='Overwrite fsaverage directories, if present')
 
 
 class BIDSFreeSurferDirOutputSpec(TraitedSpec):
@@ -234,14 +235,17 @@ class BIDSFreeSurferDir(SimpleInterface):
         make_folder(subjects_dir)
         self._results['subjects_dir'] = subjects_dir
 
-        source = os.path.join(self.inputs.freesurfer_home, 'subjects',
-                              'fsaverage')
-        dest = os.path.join(subjects_dir, 'fsaverage')
-        # Finesse is overrated. Either leave it alone or completely clobber it.
-        if os.path.exists(dest) and self.inputs.overwrite_fsaverage:
-            rmtree(dest)
-        if not os.path.exists(dest):
-            copytree(source, dest)
+        for space in self.inputs.spaces:
+            # Skip non-freesurfer spaces and fsnative
+            if not space.startswith('fsaverage'):
+                continue
+            source = os.path.join(self.inputs.freesurfer_home, 'subjects', space)
+            dest = os.path.join(subjects_dir, space)
+            # Finesse is overrated. Either leave it alone or completely clobber it.
+            if os.path.exists(dest) and self.inputs.overwrite_fsaverage:
+                rmtree(dest)
+            if not os.path.exists(dest):
+                copytree(source, dest)
 
         return runtime
